Logically Working code
Application's logic is working, But measurements of RGB and LUV needs to
be verified.

Things still to be done:
A seperate window to display the color to be measured.
Data Grid and CSV intergration.

Note:No control validations in place , and few UI elements display names
needs to be fixed.